fix(db): restore add_knowledge_graph migration + fix BIGINT FK type mismatches
Three local-deploy blockers found while running an end-to-end smoke test
with seeded data:

1. ``database/migrations/add_knowledge_graph.sql`` was never committed
   to the main tree — it only existed in agent worktrees.  Without it,
   a fresh DB has no ``knowledge_graph_concepts`` / ``knowledge_graph_edges``
   tables and every /api/kg/* endpoint returns 500.  Restored from
   the worktree copy (75 lines, schema unchanged).

2. ``add_loinc_lab_signals.sql``: ``signals_concept_id INT UNSIGNED``
   could not FK to ``knowledge_graph_concepts.id`` which is BIGINT
   UNSIGNED.  Fixed to BIGINT UNSIGNED.

3. ``add_atc_hierarchy.sql``: same problem on the ``concept_id`` column
   in the ATC indications table.  Fixed to BIGINT UNSIGNED.

Verified end-to-end after applying all migrations + running every KG seed:
  - kg_smoke_test.py: **52/52 PASS**, 0 FAIL
  - KG concepts seeded: 153 SNOMED, 134 ICD-10, 86 ATC classes, etc.
  - 65 brand→generic mappings, 121 comorbidity patterns, 251 specialty
    priors, 46 aliases, 50 evidence rules — all live.
  - /api/kg/stats returns full counts.
  - /api/kg/concepts/* and /api/kg/specialty/* all 200.

Without these fixes, a fresh local environment cannot run KG features
even after applying every committed migration in sequence.

+++ database/migrations/add_atc_hierarchy.sql
@@ -37,8 +37,8 @@ CREATE TABLE IF NOT EXISTS knowledge_graph_concepts (
 
 CREATE TABLE IF NOT EXISTS knowledge_graph_edges (
   id              INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  source_concept_id INT UNSIGNED NOT NULL,
-  target_concept_id INT UNSIGNED NOT NULL,
+  source_concept_id BIGINT UNSIGNED NOT NULL,
+  target_concept_id BIGINT UNSIGNED NOT NULL,
   relation        VARCHAR(60)  NOT NULL,           -- 'has_indication', 'is_a', 'treats', 'maps_to_hcc'
   weight          DECIMAL(5,4) DEFAULT 1.0000,
   metadata        JSON         NULL,
@@ -65,7 +65,7 @@ CREATE TABLE IF NOT EXISTS kg_atc_classes (
   name VARCHAR(255) NOT NULL,
   level TINYINT NOT NULL,                            -- 1-5
   parent_atc_code VARCHAR(20) NULL,
-  concept_id INT UNSIGNED NOT NULL,                  -- FK to knowledge_graph_concepts (ontology=atc)
+  concept_id BIGINT UNSIGNED NOT NULL,                  -- FK to knowledge_graph_concepts (ontology=atc)
   indication_concept_ids JSON NULL,                  -- [concept_id, ...] for primary indications
   is_active TINYINT(1) DEFAULT 1,
   created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
